requirements.txt
fastapi
uvicorn
sqlalchemy
databases
jinja2
python-socketio[client]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv